// frontend-shadcn-test/src/types/image.ts
export interface AutoTagsData {
  caption: string;
  taglist: string;
  rating: Record<string, number>;
  general: Record<string, number>;
  character: Record<string, number>;
  model: string;
  thresholds: {
    general: number;
    character: number;
  };
  tagged_at: string;
}